make installer more friendly fixes #6

+++ install/uninstall.sh
@@ -6,9 +6,11 @@ set -euo pipefail
 install_dir=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 marple_pst_dir=$(dirname $install_dir)
 
-source "$marple_pst_dir"/marple_pst_miniconda/bin/activate
-for i in $(seq ${CONDA_SHLVL}); do conda deactivate ; done
-conda env remove -n marple-pst
+if [ -f "$marple_pst_dir"/marple_pst_miniconda/bin/activate ] ; then
+    source "$marple_pst_dir"/marple_pst_miniconda/bin/activate
+    for i in $(seq ${CONDA_SHLVL}); do conda deactivate ; done
+    conda env remove -n marple-pst
+fi
 
-rm "$install_dir"/marple_pst_miniconda.sh
-rm -rf "$marple_pst_dir"/marple_pst_miniconda
+[ -f "$install_dir"/marple_pst_miniconda.sh ] && rm "$install_dir"/marple_pst_miniconda.sh
+[ -d "$marple_pst_dir"/marple_pst_miniconda ] && rm -rf "$marple_pst_dir"/marple_pst_miniconda
